+++ guja-gae/src/main/java/com/wadpam/guja/config/GujaGaeModule.java
@@ -23,18 +23,14 @@ package com.wadpam.guja.config;
  */
 
 import com.google.inject.AbstractModule;
-import com.google.inject.persist.PersistService;
-import com.google.inject.persist.UnitOfWork;
-import com.wadpam.guja.admintask.GAEAdminTaskQueue;
 import com.wadpam.guja.admintask.AdminTaskQueue;
+import com.wadpam.guja.admintask.GAEAdminTaskQueue;
 import com.wadpam.guja.api.GAEBlobResource;
 import com.wadpam.guja.cache.CacheBuilder;
 import com.wadpam.guja.cache.CacheBuilderProvider;
 import com.wadpam.guja.cache.MemCacheBuilderProvider;
 import com.wadpam.guja.environment.GAEServerEnvironment;
 import com.wadpam.guja.environment.ServerEnvironment;
-import net.sf.mardao.dao.DatastoreSupplier;
-import net.sf.mardao.dao.Supplier;
 
 
 /**
@@ -47,11 +43,6 @@ public class GujaGaeModule extends AbstractModule {
 
   @Override
   protected void configure() {
-
-    bind(UnitOfWork.class).to(MardaoDatastoreUnitOfWork.class);
-    bind(PersistService.class).to(MardaoDatastorePersistService.class);
-
-    bind(Supplier.class).to(DatastoreSupplier.class);
 
     bind(CacheBuilderProvider.class).to(MemCacheBuilderProvider.class);
     bind(CacheBuilder.class).toProvider(MemCacheBuilderProvider.class);
