+++ guja-gae/src/main/java/com/wadpam/guja/persist/MardaoDatastorePersistService.java
@@ -1,4 +1,4 @@
-package com.wadpam.guja.config;
+package com.wadpam.guja.persist;
 
 /*
  * #%L
